fix: move mantle pervasive graph types to graph/graph_types
