+++ graph/generate/schemagen.go
@@ -2,7 +2,7 @@ package generate
 
 import (
 	"fmt"
-	"github.com/terra-project/mantle-sdk/graph/scalars"
+	"github.com/terra-project/mantle-sdk/graph/graph_types"
 	"github.com/terra-project/mantle-sdk/serdes"
 	"reflect"
 	"strings"
@@ -133,7 +133,7 @@ func buildResponseType(t reflect.Type, tName string, parentName string) graphql.
 	}
 
 	// check if this field is scalar
-	if scalar, isScalar := scalars.IsCosmosScalar(t); isScalar {
+	if scalar, isScalar := graph_types.IsCosmosScalar(t); isScalar {
 		return scalar
 	}
