+++ graph/graph_types/mantle_pervasives.go
@@ -1,24 +1,27 @@
-package types
+package graph_types
 
 import (
 	"github.com/graphql-go/graphql"
 )
 
-var MantleScalars = []graphql.Output{
-	Order,
+type JSONScalar string // TODO: must go
+var mantlePervasives = map[string]graphql.Output{
+	"Order": Order,
+}
+
+func GetMantlePervasiveByName(typeName string) graphql.Output {
+	return mantlePervasives[typeName]
 }
 
 var Order = graphql.NewEnum(graphql.EnumConfig{
 	Name: "Order",
 	Values: graphql.EnumValueConfigMap{
 		"ASC": &graphql.EnumValueConfig{
-			Value: "ASC",
-			//DeprecationReason: "",
+			Value:       "ASC",
 			Description: "Ascending order",
 		},
 		"DESC": &graphql.EnumValueConfig{
-			Value: "DESC",
-			//DeprecationReason: "",
+			Value:       "DESC",
 			Description: "Ascending order",
 		},
 	},
